requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.2.2
contourpy==1.2.1
convert-numbers==0.4
cycler==0.12.1
Django==4.2
django-filter==24.2
django-smart-selects==1.6.0
fonttools==4.51.0
jdatetime==3.6.2
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-telegram-bot==13.7
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.11.0
tzlocal==5.2
